support function call with NO argument
When parsing "foo()", we first applied node->funcname = tok->str,
but tok->str originates from the parser p, which does not terminate tok->str with '\0'.

therefore we modified this by writing node->funcname = strndup(tok->str, tok->len).
However, strndup is not supported in this compiler,
so we implemented this code in container.c